pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gliographseg"
version = "0.1.0"
description = "Graph-based brain tumor segmentation using superpixels and GNNs"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name="Salvatore Calderaro" }
]
dependencies = [
  "torch",
  "torch-geometric",
  "numpy",
  "networkx",
  "scikit-image",
  "Pillow"
]
requires-python = ">=3.8"

[tool.setuptools]
packages = ["package.gliographseg"]

[tool.setuptools.package-data]
"package.gliographseg" = ["segmentgnn.pth"]